+++ mayan/apps/checkouts/forms.py
@@ -54,7 +54,7 @@ class DocumentCheckOutDetailForm(DetailForm):
             )
 
         kwargs['extra_fields'] = extra_fields
-        super(DocumentCheckOutDetailForm, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     class Meta:
         fields = ()