+++ HISTORY.rst
@@ -76,3 +76,4 @@
 - Add document version transformations clear and clone views.
 - Add custom API paginator. Defaults to page size of 10, maximum page size of
   100 and provides the 'page_size' query parameter.
+- Remove NewFileBlockViewTestCase and simplify the checkout logic.